Set URLs to https if HTTP X-Forwarded-Proto=https
This line changes all URLs to https URLs if the Framadate installation
is behind a reverse proxy, and the reverse proxy adds the
X-Forwarded-Proto header with the 'https' value. Configuration for
nginx:

proxy_set_header X-Forwarded-Proto https;

This setup is currently used by the Framadate installation of
Bits of Freedom (https://kies.bof.nl).