Remove the “disabled” class from fin_sondage_autre
Currently, the button still looks disabled even when more than one choice is filled.

In classic_poll.js, the disabled *property* is set/unset according to isSubmitChoicesAvalaible(). The disabled *class* is not changed after being set in the template.
In the bootstrap css, either the disabled class or the disabled gray out the button. Let’s just use the disabled property.